requirements.txt
fiftyone
torch
torchvision
numpy
pillow
scipy
matplotlib
